JsonValidator has been moved to client from core #403
